requirements.txt
google-generativeai
Pillow
numpy
pandas
tqdm
unstructured
ipython
ratelimit
opencv-python
python-dotenv